REMOVE ME: Remove hardcoded RISC-V multilibs
**THIS SHOULD BE REMOVED**
This should be a patch, but right now llvm itself isn't being patched.
So, just keep this in as a commit for now.

This might not be necessary for our final configs, but for the temp
ones I'm experimenting with, this messes up the library include paths
in a way that isn't consistent with Arm/AArch64.

We don't want these multilibs anyway, so just remove it for now.

Signed-off-by: Jonathon Penix <jpenix@quicinc.com>