Add RISC-V multilib support
This should be a decent enough start for RISC-V multilib support. The
idea here is, for each variant that we specify, match only against
supersets of the extensions for the variants we identify.

Supersets are identified naively by allowing arbitrary extra extensions
to be specified, so long as the "baseline" extensions from the variant
are expected. Given that we're looking at the canonical arch string, I
think this is sufficient and correct?

There's two maybe-questionable choices I made here:
- Specifying the canonical arch string in multilib.json seems rather
  fiddly and hard to read to me, so we allow passing the non-canonical
  form and will canonicalize it when generating the multilib YAML. I'm
  currently just doing this via `llvm-readelf -A`--I'm unsure if there
  is a simpler or more robust/correct way of doing this.
- For matching against canonical arch strings, I don't do any special
  handling of extension versions--exact matches are required. I'm not
  sure what the right approach here is--as far as I know, we don't
  support different versions of the same extensions currently and I
  generally expect the compiler/libraries/multilib.yaml file to be
  tightly coupled. Not handling extension versions also makes this
  far easier to implement, so this seems like a reasonable enough
  starting place to me.

Signed-off-by: Jonathon Penix <jpenix@quicinc.com>