WIP: Rework MUSL support to avoid using the build scripts
Looking at the variants we'll need to support eventually, it turns out
some of them have dependencies on ex: the old compiler-rt naming we
used. I don't particularly want to touch those scripts, so just add
this as an alternative. If we actually care about the `standalone` and
`uselock` variants, we can separately add those.

Signed-off-by: Jonathon Penix <jpenix@quicinc.com>